requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
